+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Changed
+ - Re-enabled keep alive by default.
+
 ## v2.0.2 - 2016-04-18
 
 ### Fixed